+++ database/migrations/2015_12_05_090741_create_human_resources_table.php
@@ -18,7 +18,7 @@ class CreateHumanResourcesTable extends Migration
 
             $table->integer('city')->unsigned();
 
-            $table->integer('population')->unsigned()->default(10);
+            $table->float('population')->unsigned()->default(10);
             $table->integer('workers')->unsigned()->default(5);
             $table->integer('settlers')->unsigned();
